+++ lib/smtp_server.dart
@@ -1,14 +1,10 @@
-import 'dart:io';
-
 export 'smtp_server/gmail.dart';
 export 'smtp_server/hotmail.dart';
 export 'smtp_server/mailgun.dart';
 export 'smtp_server/qq.dart';
 export 'smtp_server/yahoo.dart';
 
-
 class SmtpServer {
-  final String name;
   final String host;
   final int port;
   final bool ignoreBadCertificate;
@@ -16,14 +12,15 @@ class SmtpServer {
   final bool allowInsecure;
   final String username;
   final String password;
+  final String xoauth2Token;
 
   SmtpServer(this.host,
-      {this.port: 587,
+      {this.port = 587,
       String name,
-      this.ignoreBadCertificate: false,
-      this.ssl: false,
-      this.allowInsecure: false,
+      this.ignoreBadCertificate = false,
+      this.ssl = false,
+      this.allowInsecure = false,
       this.username,
-      this.password})
-      : this.name = name ?? Platform.localHostname;
+      this.password,
+      this.xoauth2Token});
 }