+++ lib/src/entities/attachment.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
+import 'dart:convert' as convert;
 import 'dart:io';
 
-import 'package:dart2_constant/convert.dart' as convert;
 import 'package:mime/mime.dart' as mime;
 import 'package:path/path.dart';
 
@@ -13,16 +13,14 @@ enum Location {
   attachment
 }
 
-/**
- * Represents a single email attachment.
- *
- * You may specify a [File], a [Stream] or just a [String] of [data].
- * [cid] allows you to specify the content id.
- *
- * When [location] is set to [Location.inline] The attachment (usually image)
- * can be referenced using:
- * `cid:yourCid`.  For instance: `<img src="cid:yourCid" />`
- */
+/// Represents a single email attachment.
+///
+/// You may specify a [File], a [Stream] or just a [String] of [data].
+/// [cid] allows you to specify the content id.
+///
+/// When [location] is set to [Location.inline] The attachment (usually image)
+/// can be referenced using:
+/// `cid:yourCid`.  For instance: `<img src="cid:yourCid" />`
 abstract class Attachment {
   String cid;
   Location location = Location.attachment;
@@ -35,14 +33,9 @@ class FileAttachment extends Attachment {
   final File _file;
 
   FileAttachment(this._file, {String contentType, String fileName}) {
-    if (contentType == null) {
-      this.contentType = mime.lookupMimeType(_file.path);
-    } else {
-      this.contentType = contentType;
-    }
-    if (this.contentType == null) {
-      this.contentType = 'application/octet-stream';
-    }
+    this.contentType = contentType ??
+        mime.lookupMimeType(_file.path) ??
+        'application/octet-stream';
     this.fileName = fileName ?? basename(_file.path);
   }
 
@@ -66,20 +59,15 @@ class StringAttachment extends Attachment {
   final String _data;
 
   StringAttachment(this._data, {String contentType, String fileName}) {
-    if (contentType == null) {
-      this.contentType = mime.lookupMimeType(fileName ?? 'abc.txt',
-          headerBytes: convert.utf8.encode(_data));
-    } else {
-      this.contentType = contentType;
-    }
-    if (this.contentType == null) {
-      this.contentType = 'text/plain';
-    }
+    this.contentType = contentType ??
+        mime.lookupMimeType(fileName ?? 'abc.txt',
+            headerBytes: convert.utf8.encode(_data)) ??
+        'text/plain';
     this.fileName = fileName;
   }
 
   @override
   // There will be only one element in the stream: the utf8 encoded string.
   Stream<List<int>> asStream() =>
-      new Stream.fromIterable([convert.utf8.encode(_data)]);
+      Stream.fromIterable([convert.utf8.encode(_data)]);
 }